WH-24; Create netcode for event endpoints
Update main to test new code

+++ src/models/Event.cpp
@@ -17,7 +17,7 @@ Event::Event(json &json):
     } else {
         endDate = json["endDate"];
     }
-    weekly = json["weekly"];
+    weekly = (int) json["weekly"];
 }
 
 uuid Event::getId() {
